mygentic/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mygentic"
version = "0.1.0"
description = "A comprehensive toolkit for agentic AI systems including web scraping, document generation, API integrations, and workflow orchestration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["ai", "agents", "web-scraping", "automation", "workflows"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core shared dependencies - always installed
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0", 
    "loguru>=0.7.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
]

# Web scraping tools
web-scraping = [
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0", 
    "firecrawl-py>=0.0.1",
    "lxml>=4.9.0",
    "scrapy>=2.11.0",
    "playwright>=1.40.0",
    "google-genai>=0.1.0",
]

# Document generation tools  
document-generation = [
    "jinja2>=3.1.0",
    "reportlab>=4.0.0",
    "pypdf>=3.17.0", 
    "markdown>=3.5.0",
    "weasyprint>=60.0",
    "python-docx>=1.1.0",
    "pdfkit>=1.0.0",
    "latex>=0.7.0",
]

# LangGraph agents
langgraph-agents = [
    "langgraph>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0", 
    "langchain-community>=0.0.1",
    "langsmith>=0.1.0",
    "tavily-python>=0.3.0",
]

# CrewAI workflows (requires Python 3.10+)
crewai-workflows = [
    "crewai>=0.28.0",
    "crewai-tools>=0.4.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
]

# MCP tools
mcp-tools = [
    "mcp>=1.0.0",
    "websockets>=12.0",
    "anyio>=4.0.0",
]

# API integrations
api-integrations = [
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "responses>=0.24.0",
]

# Install all functionality
all = [
    "mygentic[web-scraping]",
    "mygentic[document-generation]", 
    "mygentic[langgraph-agents]",
    "mygentic[crewai-workflows]",
    "mygentic[mcp-tools]",
    "mygentic[api-integrations]",
]

[project.urls]
Homepage = "https://github.com/yourusername/mygentic"
Documentation = "https://mygentic.readthedocs.io"
Repository = "https://github.com/yourusername/mygentic"
"Bug Tracker" = "https://github.com/yourusername/mygentic/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mygentic*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["mygentic"]

[tool.pytest.ini_options]
testpaths = ["tests", "*/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=mygentic --cov-report=term-missing --cov-report=html"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true